requirements.txt
asgiref==3.2.7
boto3==1.13.11
botocore==1.16.11
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.6
django-common-helpers==0.9.2
django-cron==0.5.1
django-mysql==3.5.0
django-rest-knox==4.1.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
idna==2.9
jmespath==0.10.0
mysqlclient==1.4.6
Pillow==7.1.2
pyasn1==0.4.8
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
selenium==3.141.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
